DLAPP FINAL v1.5

+++ nn_predict.py
@@ -13,15 +13,15 @@ def relu(x):
 def softmax(x):
     # TODO: Implement the SoftMax function
     # return x
-    x = np.array(x)
+    x = np.array(x, dtype=np.float64)  # 確保使用高精度
 
-    # 處理一維數組的情況
     if x.ndim == 1:
+        # 一維數組情況
         x_max = np.max(x)
         exp_x = np.exp(x - x_max)
         return exp_x / np.sum(exp_x)
     else:
-        # 處理多維數組的情況
+        # 多維數組情況
         x_max = np.max(x, axis=-1, keepdims=True)
         exp_x = np.exp(x - x_max)
         return exp_x / np.sum(exp_x, axis=-1, keepdims=True)
